Remove RelocatedNotEstablished trait
